+++ src/main/java/com/ae2powertools/CommonProxy.java
@@ -1,6 +1,8 @@
 package com.ae2powertools;
 
 import net.minecraftforge.common.MinecraftForge;
+import net.minecraftforge.fml.common.Loader;
+import net.minecraftforge.fml.common.Optional;
 import net.minecraftforge.fml.common.event.FMLInitializationEvent;
 import net.minecraftforge.fml.common.event.FMLPreInitializationEvent;
 
@@ -15,12 +17,22 @@ import com.ae2powertools.features.tuner.PriorityTunerEventHandler;
  */
 public class CommonProxy {
 
+    private static final String TOP_MODID = "theoneprobe";
+
     public void preInit(FMLPreInitializationEvent event) {
+        // Register TOP integration (must be on common, even though TOP is mostly client-side)
+        if (Loader.isModLoaded(TOP_MODID)) registerTheOneProbeIntegration();
+
         MinecraftForge.EVENT_BUS.register(new PriorityTunerEventHandler());
         MinecraftForge.EVENT_BUS.register(new LevelMonitorAlarmEventHandler());
     }
 
     public void init(FMLInitializationEvent event) {
         AEApi.instance().registries().wireless().registerWirelessHandler(ItemRegistry.REMOTE_STORAGE_MONITOR);
+    }
+
+    @Optional.Method(modid = TOP_MODID)
+    private static void registerTheOneProbeIntegration() {
+        com.ae2powertools.integration.theoneprobe.PowerToolsTheOneProbePlugin.register();
     }
 }